implemented more tests

+++ src/main/java/ch/uzh/ifi/hase/soprafs23/Betting/Instruction.java
@@ -17,7 +17,7 @@ public class Instruction {
     @Column(name = "number")
     private double number;
 
-    public Instruction(Long ownerID, InstructionType type, int number){
+    public Instruction(Long ownerID, InstructionType type, double number){
         this.number = number;
         this.ownerID = ownerID;
         this.type = type;
